requirements.txt
python-telegram-bot~=20.0a4
pandas~=1.4.4
pdfkit~=1.0.0
python-dotenv~=0.21.0
psycopg2~=2.9.3
